fix: add reusable activiti-cloud-service-common-liquibase module
