+++ activiti-cloud-audit-service/activiti-cloud-services-audit-liquibase/pom.xml
@@ -10,8 +10,8 @@
   
   <dependencies>
     <dependency>
-      <groupId>org.liquibase</groupId>
-      <artifactId>liquibase-core</artifactId>
+      <groupId>org.activiti.cloud</groupId>
+      <artifactId>activiti-cloud-service-common-liquibase</artifactId>
     </dependency>
     <dependency>
       <groupId>org.springframework.boot</groupId>
