+++ activiti-cloud-audit-service/activiti-cloud-services-audit-liquibase/src/main/java/org/activiti/cloud/services/audit/liquibase/ActivitiCloudAuditLiquibaseAutoConfiguration.java
@@ -17,15 +17,14 @@ package org.activiti.cloud.services.audit.liquibase;
 
 import javax.sql.DataSource;
 
+import org.activiti.cloud.common.liquibase.SpringLiquibaseConfigurationSupport;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.autoconfigure.liquibase.LiquibaseProperties;
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.ResourceLoaderAware;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
-import org.springframework.core.io.ResourceLoader;
 
 import liquibase.exception.LiquibaseException;
 import liquibase.integration.spring.SpringLiquibase;
@@ -33,14 +32,7 @@ import liquibase.integration.spring.SpringLiquibase;
 @Configuration
 @ConditionalOnProperty(prefix = "spring.liquibase", name = "enabled", matchIfMissing = true)
 @PropertySource("classpath:config/audit-liquibase.properties")
-public class ActivitiCloudAuditLiquibaseAutoConfiguration implements ResourceLoaderAware {
-
-    private ResourceLoader resourceLoader;
-
-    @Override
-    public void setResourceLoader(ResourceLoader resourceLoader) {
-        this.resourceLoader = resourceLoader;
-    }
+public class ActivitiCloudAuditLiquibaseAutoConfiguration extends SpringLiquibaseConfigurationSupport {
 
     @Bean
     @ConditionalOnMissingBean(name = "auditLiquibase")
@@ -52,24 +44,5 @@ public class ActivitiCloudAuditLiquibaseAutoConfiguration implements ResourceLoa
     @ConfigurationProperties(prefix = "spring.audit.liquibase")
     public LiquibaseProperties auditLiquibaseProperties() {
         return new LiquibaseProperties();
-    }
-
-    private SpringLiquibase buildSpringLiquibase(DataSource dataSource, LiquibaseProperties properties) {
-        SpringLiquibase liquibase = new SpringLiquibase();
-        liquibase.setDataSource(dataSource);
-        liquibase.setResourceLoader(resourceLoader);
-        liquibase.setChangeLog(properties.getChangeLog());
-        liquibase.setContexts(properties.getContexts());
-        liquibase.setDefaultSchema(properties.getDefaultSchema());
-        liquibase.setLiquibaseTablespace(properties.getLiquibaseTablespace());
-        liquibase.setDropFirst(properties.isDropFirst());
-        liquibase.setShouldRun(properties.isEnabled());
-        liquibase.setLabels(properties.getLabels());
-        liquibase.setChangeLogParameters(properties.getParameters());
-        liquibase.setRollbackFile(properties.getRollbackFile());
-        liquibase.setDatabaseChangeLogTable(properties.getDatabaseChangeLogTable());
-        liquibase.setDatabaseChangeLogLockTable(properties.getDatabaseChangeLogLockTable());
-        liquibase.setTestRollbackOnUpdate(properties.isTestRollbackOnUpdate());
-        return liquibase;
     }
 }
