+++ activiti-cloud-query-service/activiti-cloud-services-query/activiti-cloud-services-query-liquibase/src/main/java/org/activiti/cloud/services/query/liquibase/ActivitiCloudQueryLiquibaseAutoConfiguration.java
@@ -17,15 +17,14 @@ package org.activiti.cloud.services.query.liquibase;
 
 import javax.sql.DataSource;
 
+import org.activiti.cloud.common.liquibase.SpringLiquibaseConfigurationSupport;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.autoconfigure.liquibase.LiquibaseProperties;
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.ResourceLoaderAware;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
-import org.springframework.core.io.ResourceLoader;
 
 import liquibase.exception.LiquibaseException;
 import liquibase.integration.spring.SpringLiquibase;
@@ -33,14 +32,7 @@ import liquibase.integration.spring.SpringLiquibase;
 @Configuration
 @ConditionalOnProperty(prefix = "spring.liquibase", name = "enabled", matchIfMissing = true)
 @PropertySource("classpath:config/query-liquibase.properties")
-public class ActivitiCloudQueryLiquibaseAutoConfiguration  implements ResourceLoaderAware {
-
-    private ResourceLoader resourceLoader;
-
-    @Override
-    public void setResourceLoader(ResourceLoader resourceLoader) {
-        this.resourceLoader = resourceLoader;
-    }
+public class ActivitiCloudQueryLiquibaseAutoConfiguration extends SpringLiquibaseConfigurationSupport {
 
     @Bean
     @ConditionalOnMissingBean(name = "queryLiquibase")
@@ -52,24 +44,5 @@ public class ActivitiCloudQueryLiquibaseAutoConfiguration  implements ResourceLo
     @ConfigurationProperties(prefix = "spring.query.liquibase")
     public LiquibaseProperties queryLiquibaseProperties() {
         return new LiquibaseProperties();
-    }
-
-    private SpringLiquibase buildSpringLiquibase(DataSource dataSource, LiquibaseProperties properties) {
-        SpringLiquibase liquibase = new SpringLiquibase();
-        liquibase.setDataSource(dataSource);
-        liquibase.setResourceLoader(resourceLoader);
-        liquibase.setChangeLog(properties.getChangeLog());
-        liquibase.setContexts(properties.getContexts());
-        liquibase.setDefaultSchema(properties.getDefaultSchema());
-        liquibase.setLiquibaseTablespace(properties.getLiquibaseTablespace());
-        liquibase.setDropFirst(properties.isDropFirst());
-        liquibase.setShouldRun(properties.isEnabled());
-        liquibase.setLabels(properties.getLabels());
-        liquibase.setChangeLogParameters(properties.getParameters());
-        liquibase.setRollbackFile(properties.getRollbackFile());
-        liquibase.setDatabaseChangeLogTable(properties.getDatabaseChangeLogTable());
-        liquibase.setDatabaseChangeLogLockTable(properties.getDatabaseChangeLogLockTable());
-        liquibase.setTestRollbackOnUpdate(properties.isTestRollbackOnUpdate());
-        return liquibase;
     }
 }
