+++ activiti-cloud-service-common/activiti-cloud-service-common-dependencies/pom.xml
@@ -102,6 +102,11 @@
         <artifactId>activiti-cloud-service-error-handlers</artifactId>
         <version>${project.version}</version>
       </dependency>
+      <dependency>
+        <groupId>org.activiti.cloud</groupId>
+        <artifactId>activiti-cloud-service-common-liquibase</artifactId>
+        <version>${project.version}</version>
+      </dependency>
     </dependencies>
   </dependencyManagement>
