+++ activiti-cloud-service-common/activiti-cloud-service-common-liquibase/pom.xml
@@ -1,0 +1,21 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <artifactId>activiti-cloud-service-common-dependencies</artifactId>
+    <groupId>org.activiti.cloud</groupId>
+    <version>7.1.0-SNAPSHOT</version>
+    <relativePath>../activiti-cloud-service-common-dependencies</relativePath>
+  </parent>
+  <artifactId>activiti-cloud-service-common-liquibase</artifactId>
+  <dependencies>
+    <dependency>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-autoconfigure</artifactId>
+      <optional>true</optional>
+    </dependency>      
+    <dependency>
+      <groupId>org.liquibase</groupId>
+      <artifactId>liquibase-core</artifactId>
+    </dependency>  
+  </dependencies>
+</project>
