+++ activiti-cloud-service-common/pom.xml
@@ -31,6 +31,7 @@
     <module>activiti-cloud-service-error-handlers</module>
     <module>activiti-cloud-services-test-security-keycloak</module>
     <module>activiti-cloud-services-test-containers</module>
+    <module>activiti-cloud-service-common-liquibase</module>
   </modules>
   <properties>
     <activiti.version>7.1.279</activiti.version>
